Only install flatpaks if necessary

+++ scripts/setup_flatpak.sh
@@ -25,9 +25,13 @@ flatpak remote-add --if-not-exists --user gnome-nightly "$SCRIPT_DIR/../assets/f
 # flatpak remote-add --user --if-not-exists kde "$SCRIPT_DIR/../assets/kderuntime.flatpakrepo"
 # flatpak remote-add --user --if-not-exists freedesktop-sdk https://releases.freedesktop-sdk.io/freedesktop-sdk.flatpakrepo
 
-# Version that includes the ssh-add CLI command
-flatpak install --user https://dl.flathub.org/build-repo/84660/org.keepassxc.KeePassXC.flatpakref
-# flatpak install --user flathub org.keepass.KeePassXC
+if flatpak list --app | grep org.keepassxc.KeePassXC &> /dev/null; then
+    echo "KeePassXC is already installed."
+else
+    # Version that includes the ssh-add CLI command
+    flatpak install --user https://dl.flathub.org/build-repo/84660/org.keepassxc.KeePassXC.flatpakref
+    # flatpak install --user flathub org.keepass.KeePassXC
+fi
 
 KPXC_CONFIG_DIR="$HOME/.config/keepassxc"
 KPXC_CONFIG_PATH="$KPXC_CONFIG_DIR/keepassxc.ini"
@@ -36,7 +40,11 @@ if [[ ! -d "$KPXC_CONFIG_DIR" ]]; then
 fi
 cp "$SCRIPT_DIR/../assets/keepassxc/keepassxc.ini" "$KPXC_CONFIG_PATH"
 
-flatpak install --user flathub org.gnome.Evince
+if flatpak list --app | grep org.gnome.Evince &> /dev/null; then
+    echo "Evince is already installed."
+else
+    flatpak install --user flathub org.gnome.Evince
+fi
 
 # flatpak install --user flathub io.neovim.nvim
